중간커밋

+++ src/main/java/com/example/todo/service/TodoSvcImpl.java
@@ -1,16 +1,16 @@
 package com.example.todo.service;
 
+import com.example.todo.dto.request.TodoDeleteRequestDto;
 import com.example.todo.dto.response.ResponseTodoDto;
 import com.example.todo.dto.request.TodoCreateRequestDto;
 import com.example.todo.dto.request.TodoUpdateRequestDto;
 import com.example.todo.entity.Todo;
 import com.example.todo.entity.User;
-import com.example.todo.exception.DbException;
-import com.example.todo.exception.FailToCreateTodoException;
-import com.example.todo.exception.TodoNotFoundException;
+import com.example.todo.exception.*;
 import com.example.todo.repository.TodoRepository;
 import com.example.todo.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,9 +35,9 @@ public class TodoSvcImpl implements TodoService {
 
         try {
             all = todoRepo.findAll(userId, page);
-        } catch (DbException e) {
+        } catch (DataAccessException e) {
             log.error(e.getMessage(),e);
-            throw new TodoNotFoundException("TodoNotFoundException");
+            throw new TodoNotFoundException("일정을 조회할수없습니다.");
         }
 
         //List<Entity> -> List<Dto>
@@ -55,7 +55,16 @@ public class TodoSvcImpl implements TodoService {
     }
 
     @Override
-    public int deleteById(Long id) {
+    public int deleteById(Long id, TodoDeleteRequestDto requestDto) {
+        Todo byId = todoRepo.findById(id);
+
+        if(byId.getUser().getId().equals(requestDto.getUserId())) {
+            throw new AccessDeniedException("작성자 아님");
+        }
+        if(!byId.getPwd().equals(requestDto.getPwd())) {
+            throw new InvalidPasswordException();
+        }
+
         return todoRepo.deleteById(id);
     }
 
@@ -67,7 +76,7 @@ public class TodoSvcImpl implements TodoService {
 
         try {
             user = userRepo.findById(todoDto.getUserId());
-        } catch (DbException e) {
+        } catch (DataAccessException e) {
             log.error(e.getMessage(),e);
             throw new FailToCreateTodoException("User with id " + todoDto.getUserId() + " not found");
         }
@@ -76,7 +85,7 @@ public class TodoSvcImpl implements TodoService {
         Todo insert;
         try {
             insert = todoRepo.insert(todo);
-        } catch (DbException e) {
+        } catch (DataAccessException e) {
             log.error(e.getMessage(),e);
             throw new FailToCreateTodoException("fail to create todo");
         }
@@ -88,7 +97,17 @@ public class TodoSvcImpl implements TodoService {
     }
 
     @Override
+    @Transactional
     public int update(Long id, TodoUpdateRequestDto todoDto) {
+        Todo byId = todoRepo.findById(id);
+
+        if(!byId.getUser().getId().equals(todoDto.getUserId())) {
+            throw new AccessDeniedException("작성자 아님");
+        }
+        if(!byId.getPwd().equals(todoDto.getPwd())) {
+            throw new InvalidPasswordException();
+        }
+
         return todoRepo.update(id, todoDto);
     }
 }
